DWA9 - begin refactor
Fixed structure and moved files around for better readability
